+++ ext/FlexiChainsMakieExt/traceplot.jl
@@ -3,20 +3,20 @@ function _default_traceplot_axis(k::FC.ParameterOrExtra)
 end
 
 """
-    FlexiChains.mtraceplot(
+    FlexiChains.Makie.traceplot(
         chn::FC.FlexiChain[, param_or_params];
         kwargs...,
     )
 
 Create trace plots for the specified parameters in the chain.
 
-$(FC._PARAM_DOCSTRING("FlexiChains.mtraceplot"))
+$(FC.PlotUtils._PARAM_DOCSTRING("FlexiChains.Makie.traceplot"))
 
 # Keyword arguments
 
 $(MAKIE_KWARGS_DOCSTRING)
 """
-function FC.mtraceplot(
+function FC.Makie.traceplot(
         chn::FC.FlexiChain,
         param_or_params = FC.Parameter.(FC.parameters(chn));
         layout::Union{Tuple{Int, Int}, Nothing} = nothing,
@@ -34,7 +34,7 @@ function FC.mtraceplot(
     # This order means that plots go from left to right before going to the next row
     indices = Iterators.product(1:ncols, 1:nrows)
     for ((col, row), k) in zip(indices, keys_to_plot)
-        a, p = FC.mtraceplot!(
+        a, p = FC.Makie.traceplot!(
             Makie.Axis(figure[row, col]; _default_traceplot_axis(k)..., axis...),
             FC.PlotUtils.FlexiChainTrace(chn, k);
             kwargs...,
@@ -49,29 +49,29 @@ end
 ########################
 # Single axis plotting #
 ########################
-function FC.mtraceplot(grid::MakieGrids, chn::FC.FlexiChain, param; axis = (;), kwargs...)
+function FC.Makie.traceplot(grid::MakieGrids, chn::FC.FlexiChain, param; axis = (;), kwargs...)
     # TODO: Error if there is already something at the grid position?
     # See e.g. https://github.com/rafaqz/DimensionalData.jl/blob/6db30de4b2e1fc7f8611b7e1dc3f89dc02c78598/ext/DimensionalDataMakieExt.jl#L85-L96
     chn = FC.PlotUtils.subset_and_split_chain(chn, param)
     k = only(keys(chn))
-    return FC.mtraceplot!(
+    return FC.Makie.traceplot!(
         Makie.Axis(grid; _default_traceplot_axis(k)..., axis...), chn, param; kwargs...
     )
 end
-function FC.mtraceplot!(ax::Makie.Axis, chn::FC.FlexiChain, param; kwargs...)
+function FC.Makie.traceplot!(ax::Makie.Axis, chn::FC.FlexiChain, param; kwargs...)
     chn = FC.PlotUtils.subset_and_split_chain(chn, param)
     k = only(keys(chn))
-    a, p = FC.mtraceplot!(ax, FC.PlotUtils.FlexiChainTrace(chn, k); kwargs...)
+    a, p = FC.Makie.traceplot!(ax, FC.PlotUtils.FlexiChainTrace(chn, k); kwargs...)
     return Makie.AxisPlot(a, p)
 end
-function FC.mtraceplot!(chn::FC.FlexiChain, param; kwargs...)
-    return FC.mtraceplot!(Makie.current_axis(), chn, param; kwargs...)
+function FC.Makie.traceplot!(chn::FC.FlexiChain, param; kwargs...)
+    return FC.Makie.traceplot!(Makie.current_axis(), chn, param; kwargs...)
 end
 
 """
 This is the actual function that does the density plotting.
 """
-function FC.mtraceplot!(ax::Makie.Axis, d::FC.PlotUtils.FlexiChainTrace; kwargs...)
+function FC.Makie.traceplot!(ax::Makie.Axis, d::FC.PlotUtils.FlexiChainTrace; kwargs...)
     x = FC.iter_indices(d.chn)
     y = FC._get_raw_data(d.chn, d.param)
     nchains = size(y, 2)